requirements.txt
# Django packages
Django==4.0.2
django-import-export==2.7.1

# LINTERS
black==21.9b0
pre-commit==2.15.0


# Tailwind packages
django-tailwind==3.1.1
django-crispy-forms==1.11.2
crispy-tailwind==0.5.0


# UTILS
Pillow==8.4.0
openpyxl~=3.0.7 # excel
xlrd==1.2.0 # excel
factory-boy==3.2.1
django-debug-toolbar==3.2.4


# HEROKU DEPLOYMENT
django-environ==0.8.1
#dj-database-url==0.3.0
#dj-static==0.0.6
gunicorn==19.6.0
python-decouple==3
whitenoise
psycopg2


# TEST
selenium==4.1.2
webdriver-manager==3.5.2
